backend/api-server/api-server.requirements.txt
fastapi==0.68.0
sqlalchemy==1.3.17
psycopg2==2.8.5
databases[postgresql]==0.3.2
async-exit-stack
async-generator
rq==1.9.0
pyjwt==1.0.0
requests==2.22.0
python-multipart==0.0.5
